+++ io/features/eu.esdihumboldt.hale.io.feature.core/feature.xml
@@ -371,4 +371,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="eu.esdihumboldt.hale.io.topojson"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
