+++ io/plugins/eu.esdihumboldt.hale.io.json.test/META-INF/MANIFEST.MF
@@ -15,10 +15,6 @@ Require-Bundle: eu.esdihumboldt.hale.io.json;bundle-version="2.9.0",
  eu.esdihumboldt.cst.functions.geometric;bundle-version="4.2.0",
  org.locationtech.jts.jts-core;bundle-version="1.16.0",
  org.junit;bundle-version="4.13.0",
- eu.esdihumboldt.hale.common.instance;bundle-version="4.2.0",
- org.opengis;bundle-version="21.0.0",
- eu.esdihumboldt.util;bundle-version="4.2.0",
- org.eclipse.core.runtime;bundle-version="3.17.100",
  eu.esdihumboldt.hale.common.align.groovy;bundle-version="2.9.0"
 Import-Package: de.fhg.igd.slf4jplus,
  eu.esdihumboldt.hale.common.instance.groovy,
