+++ io/plugins/eu.esdihumboldt.hale.io.json/plugin.xml
@@ -42,14 +42,6 @@
                ref="eu.esdihumboldt.hale.io.json.gzip">
          </contentType>
       </provider>
-      <provider
-            class="eu.esdihumboldt.hale.io.json.TopoJsonInstanceWriter"
-            id="eu.esdihumboldt.hale.io.topojson.writer"
-            name="TopoJSON">
-         <contentType
-               ref="eu.esdihumboldt.hale.io.json">
-         </contentType>
-      </provider>
    </extension>
    <extension
          point="eu.esdihumboldt.hale.io.complexvalue">
