+++ io/plugins/eu.esdihumboldt.hale.io.topojson.test/.settings/org.eclipse.jdt.groovy.core.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+groovy.compiler.level=24
