+++ io/plugins/eu.esdihumboldt.hale.io.topojson/.settings/org.eclipse.jdt.launching.prefs
@@ -1,0 +1,4 @@
+#Created from default preferences Oct 21, 2022 2:28:43 PM
+#Fri Oct 21 14:28:43 CEST 2022
+eclipse.preferences.version=1
+org.eclipse.jdt.launching.PREF_STRICTLY_COMPATIBLE_JRE_NOT_AVAILABLE=ignore
