+++ io/plugins/eu.esdihumboldt.hale.io.topojson/META-INF/MANIFEST.MF
@@ -1,0 +1,18 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Topojson Support
+Bundle-SymbolicName: eu.esdihumboldt.hale.io.topojson;singleton:=true
+Bundle-Version: 4.2.0.qualifier
+Bundle-Vendor: wetransform GmbH
+Automatic-Module-Name: eu.esdihumboldt.hale.io.topojson
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: eu.esdihumboldt.hale.io.shp;bundle-version="4.2.0",
+ com.github.wetransform-os.topojson-j;bundle-version="1.0.0",
+ eu.esdihumboldt.hale.common.instance;bundle-version="4.2.0",
+ eu.esdihumboldt.hale.common.core;bundle-version="4.2.0",
+ eu.esdihumboldt.hale.common.schema;bundle-version="4.2.0",
+ org.eclipse.core.runtime;bundle-version="3.17.100",
+ org.opengis;bundle-version="21.0.0",
+ eu.esdihumboldt.util;bundle-version="4.2.0",
+ eu.esdihumboldt.hale.io.json;bundle-version="4.2.0"
+Export-Package: eu.esdihumboldt.hale.io.topojson
