+++ io/plugins/eu.esdihumboldt.hale.io.topojson/plugin.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="eu.esdihumboldt.hale.io.provider">
+      <provider
+            allowDuplicate="false"
+            class="eu.esdihumboldt.hale.io.topojson.TopoJsonInstanceWriter"
+            id="eu.esdihumboldt.hale.io.topojson.writer"
+            name="TopoJSON">
+         <contentType
+               ref="eu.esdihumboldt.hale.io.json">
+         </contentType>
+      </provider>
+   </extension>
+
+</plugin>
